requirements.txt
selenium~=4.22.0
behave~=1.2.6
allure-behave~=2.13.5